+++ 404.php
@@ -3,7 +3,7 @@
  * The template for displaying 404 pages (not found)
  *
  * @author      David Voglgsang
- * @version     1.1.2
+ * @version     1.2.2
 */
 
 get_header();
@@ -18,7 +18,14 @@ get_header();
   <article>
     <h1 class="post-title"><?php echo _e('404 Error', 'devTheme'); ?></h1>
     <p><?php _e('Page could not be found.', 'devTheme'); ?></p>
-    <a href="<?php echo get_bloginfo('url'); ?>" class="button"><?php _e('Back to home', 'devTheme'); ?></a>
+    <?php
+      if(prefix_template::$template_404_searchForm == 1):
+        echo get_search_form();
+      endif;
+      if(prefix_template::$template_404_backToHome == 1):
+        echo '<a href="' . get_bloginfo('url') . '" class="button">' . __('Back to home', 'devTheme') . '</a>';
+      endif;
+    ?>
   </article>
 </section>
 <?php get_footer() ?>