+++ classes/prefix_WPgutenberg/blocks/core-extensions/gallery/index.js
@@ -42,6 +42,7 @@
         addBulletNav,
         addPopUp,
         addPopUpPreview,
+        addDownloadButton,
       } = attributes;
       return (
         <Fragment>