+++ classes/prefix_WPgutenberg/blocks/core-extensions/image/attributes.js
@@ -2,6 +2,10 @@ const attributes = {
   addPopUp: {
     type: 'boolean',
     default: false,
+  },
+  addDownloadButton: {
+    type: 'boolean',
+    default: false,
   }
 }