+++ classes/prefix_WPgutenberg/blocks/core-extensions/image/index.js
@@ -39,6 +39,7 @@
       } = props;
       const {
         addPopUp,
+        addDownloadButton,
       } = attributes;
       return (
         <Fragment>